Adds leading 0 for single digit Anime searches
When requesting an episode of Anime with a single digit episode number,
the results would get cluttered with episodes that weren't requested.
Nzbdrone correctly ignored those episodes, but sometimes more than 100 of
those results would appear which would keep requested episodes from showing
up in search results. This change makes it so that at least 2 digits are
used in the query string when searching for Anime episodes. Besides being
a closer match to what the scene actually releases, it also cuts down on the
clutter from non-requested episodes containing the requested episode number
in the hash/resolution.